+++ lib/benchee/benchmark.ex
@@ -163,8 +163,7 @@ defmodule Benchee.Benchmark do
   end
 
   @fast_warning """
-  Warning: The function you are trying to benchmark is super fast, making time measures unreliable!
-  Benchee won't measure individual runs but rather run it a couple of times and report the average back. Measures will still be correct, but the overhead of running it n times goes into the measurement. Also statistical results aren't as good, as they are based on averages now. If possible, increase the input size so that an individual run takes more than #{@minimum_execution_time}μs
+  Warning: The function you are trying to benchmark is super fast, making measures more unreliable! See: https://github.com/PragTob/benchee/wiki/Benchee-Warnings#fast-execution-warning
   """
   defp print_fast_warning do
     IO.puts @fast_warning